Chore: clean 4 pre-existing clippy 1.94 lints
- `core/src/api/origin.rs:11-14` — doc list continuation lines
  indented to match list-item start (4 spaces) instead of
  18-space alignment that triggered `doc_overindented_list_items`
- `core/src/socket_listener.rs:1691` — drop `as u128` cast
  (`.as_nanos()` already returns u128); was
  `unnecessary_cast`
- `core/tests/db_test.rs:54` — annotate `seed_message` test
  helper with `#[allow(clippy::too_many_arguments)]`; 8-arg
  signature is a test seed pattern, not a code-smell to refactor

Pre-existing from clippy 1.94 strictness bump (introduced
mid-M3/M7 work; not caught earlier because clippy isn't a CI
gate — check.yml only runs `cargo check` + `cargo test`).
Fixing here as M9 prep cleanup; `cargo clippy --workspace
--all-targets -- -D warnings` now passes clean. cargo test
180/180 unchanged.

+++ core/src/socket_listener.rs
@@ -1688,7 +1688,7 @@ fn mint_project_id() -> String {
     use std::time::{SystemTime, UNIX_EPOCH};
     let ts = SystemTime::now()
         .duration_since(UNIX_EPOCH)
-        .map(|d| d.as_nanos() as u128)
+        .map(|d| d.as_nanos())
         .unwrap_or(0);
     let mut x: u128 = ts;
     // Splitmix-ish stir so two ids minted in the same ns differ.
